Day 6, Part 2
Detect ending index of first n unique characters in a stream.
Was passing in a reader, which would prevent redundant reads, but
it only works if subsequent calls are for >n length substrings,
so just did the redundant call. Could be improved to be a bit more
performant, but it's fine.

https://adventofcode.com/2022/day/6#part2